[issue-150] Starting new survey page
