--- cocoon/frontend/src/userAuth/survey/survey.css
@@ -1,5 +1,0 @@
-.Dotted_box {
-    outline-style: dotted;
-    margin-top: 20px;
-    margin-bottom: 20px;
-}
