+++ cocoon/frontend/src/userAuth/survey/surveySmall.css
@@ -1,0 +1,15 @@
+.survey-div {
+}
+
+.survey-box {
+    position: absolute;
+left: 0%;
+right: 0%;
+top: 0%;
+bottom: 0%;
+
+background: #F6F6F6;
+border: 1px solid rgba(0, 0, 0, 0.5);
+box-shadow: 0px 4px 4px rgba(0, 0, 0, 0.25);
+border-radius: 33px;
+}
