+++ cocoon/frontend/src/userAuth/survey/surveys.css
@@ -1,0 +1,72 @@
+.Dotted_box {
+    outline-style: dotted;
+    margin-top: 20px;
+    margin-bottom: 20px;
+}
+
+/* Roommate Groups */
+.my-survey-header {
+    /*position: absolute;*/
+    /*width: 408px;*/
+    /*height: 58px;*/
+    /*left: 516px;*/
+    /*top: 112px;*/
+
+    margin-top: 10px;
+
+    font-family: Montserrat;
+    font-style: normal;
+    font-weight: 800;
+    line-height: normal;
+    font-size: 36px;
+    text-align: center;
+
+    color: #000000;
+}
+
+.help-button {
+    margin-top: 10px;
+}
+
+.search-bar-div {
+    width: 892px;
+    height: 50px;
+}
+
+.search-button {
+    position: absolute;
+    left: 85.2%;
+    right: 0%;
+    top: 3.03%;
+    bottom: 0%;
+
+    border: 1px solid #000000;
+    box-shadow: inset 0px 4px 4px rgba(0, 0, 0, 0.25);
+}
+
+.search-bar {
+    width: 86.2%;
+    /*right: 14.8%;*/
+    position: absolute;
+    left: 0%;
+    top: 3.03%;
+    bottom: 0%;
+
+    border: 1px solid #000000;
+    box-shadow: inset 0px 4px 4px rgba(0, 0, 0, 0.25);
+    border-radius: 6px 0px 0px 6px;
+}
+
+.surveys {
+    margin-top: 10px;
+}
+
+.survey {
+    margin-top: 15px;
+    width: 399px;
+    height: 295px;
+}
+
+.survey-gap {
+    width: 3%;
+}
