+++ cocoon/userAuth/templates/userAuth/surveys.html
@@ -2,6 +2,10 @@
 {% load static %}
 {% block title %}My Surveys{% endblock %}
 
+{% block container_type %}
+    container-fluid
+{% endblock %}
+
 {% block content %}
     <div id="reactWrapper"></div>
 {% endblock %}
